+++ CHANGELOG.md
@@ -5,10 +5,23 @@
 - **Workspace Snapshot Sharing** Workspaces can now be shared as snapshots. This allows for sharing a workspace with a specific configuration, including all tabs, settings, and data. Snapshots can be shared with other users or saved for later use. Snapshots are read-only and cannot be edited after being created, but can be duplicated.
   - All sub-items (expressions, ROIs, quants, etc.) are (attempted to be) shared on snapshot-share, but if the user doesn't have edit permissions for a sub-item, they will not be able to update its sharing settings and this shows up as a warning. This sharing issue can be resolved either before or after sharing a workspace and workspaces/all sub-items can be "re-shared" later.
 - **Workspace Duplicating** Workspaces can now be duplicated. This allows for creating a copy of a workspace with all tabs, settings, and data. Duplicated workspaces can be edited and saved as new workspaces.
+- **Default Quant Selection** When creating a new workspace, a default quant is now selected (workspace configuration is also automatically saved on quant change). Default quant selection is made by the following priority hierarchy:
+  - AutoQuant PIXL (A/B)
+  - AutoQuant PIXL (Combined)
+  - AutoQuant PDS (A/B)
+  - AutoQuant PDS (Combined)
+  - < Any default imported quant >
+  - First user-created A/B quant
+  - First user-created Combined quant
+- **New Workspace Improvements** Clicking the "New Workspace" button now opens a dialog to configure the workspace name, description, tags, and scans before creation. Additionally, holding "Cmd" (or "Ctrl" on Windows) and clicking scans will now add default select them for the new workspace.
+- **Workspace Searching Improvements** Workspace search now includes searching by snapshots and creator names. Also, you can click on column names to sort the workspace list. Lastly, your "Workspaces" or "Datasets" selection will be remembered when you navigate back to the workspace list.
 - **Backup & Restore** For PIXLISE Devs/Admins to be able to back up all PIXLISE data and restore onto another environment in a more automated way.
 
 ### Bug Fixes
 
+- Fixes logout bug when clicked from the public site (redirects to wrong URL)
+- Fixes bug when plotting sigma values on Parallel Coordinates Plot with ratio columns showing
+- Fixes bug when creating new ROI from the ROI sidepanel instead of the selection dropdown (footprint was being selected)
 - Fixed intermittent connection error 18798 when UI was reconnecting to server
 
 ## 4.40.0 (2024-09-05)